+++ src/briefy/alexandria/models/__init__.py
@@ -2,6 +2,9 @@
 from briefy.alexandria.models.asset import Asset  # noQA
 from briefy.alexandria.models.base import AssetsCollection  # noQA
 from briefy.alexandria.models.collection import Collection  # noQA
+from briefy.ws.listeners import register_workflow_context_listeners
+
+import sqlalchemy as sa
 
 
 ALL_MODELS = [
@@ -9,3 +12,8 @@ ALL_MODELS = [
     AssetsCollection,
     Collection
 ]
+
+# register sqlalchemy workflow context event handlers
+register_workflow_context_listeners(ALL_MODELS)
+
+sa.orm.configure_mappers()